pyproject.toml
[project]
name = "locationsmap"
version = "0.1.0"
description = ""
authors = [
    {name = "Shapka Oleksii",email = "shapkaoleksiiu@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-filter (>=25.1,<26.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)"
]

[tool.poetry]
packages = [{include = "locationsmap", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
